+++ test/binaryen.js/global.js
@@ -13,6 +13,7 @@ function cleanInfo(info) {
 }
 
 var module = new Binaryen.Module();
+module.setFeatures(Binaryen.Features.MVP | Binaryen.Features.MutableGlobals);
 
 var initExpr = module.i32.const(1);
 var global = module.addGlobal("a-global", Binaryen.i32, false, initExpr);
@@ -27,7 +28,8 @@ console.log("getExpressionInfo(init)=" + JSON.stringify(cleanInfo(initExpInfo)))
 console.log(Binaryen.emitText(globalInfo.init));
 
 module.addGlobalExport("a-global", "a-global-exp");
-module.addGlobalImport("a-global-imp", "module", "base", Binaryen.i32);
+module.addGlobalImport("a-global-imp", "module", "base", Binaryen.i32, false);
+module.addGlobalImport("a-mut-global-imp", "module", "base", Binaryen.i32, true);
 
 assert(module.validate());
 console.log(module.emitText());