+++ CHANGELOG.md
@@ -20,8 +20,11 @@ Current Trunk
   - `i32.wait` / `i64.wait` -> `i32.atomic.wait` / `i64.atomic.wait`
 - Binaryen.js: `flags` argument in `setMemory` function is removed.
 - `atomic.fence` instruction support is added.
-- wasm-emscripten-finalize: Don't realy on name section being present in the
+- wasm-emscripten-finalize: Don't rely on name section being present in the
   input. Use the exported names for things instead.
+- Added `mutable` parameter to BinaryenAddGlobalImport.
+- Replace BinaryenSIMDBitselect* with BinaryenSIMDTernary* in the C API and add
+  qfma/qfms instructions.
 
 v88
 ---